Be more careful about the order in which we kill processes.
In particular, don't have the watchdog kill itself. Even though
perl's `kill` takes a list of pids, they actually get signalled
one at a time, and so if the watchdog is in the middle of the
list not everything will get killed like it should be. So have
the watchdog filter itself out of the list, and then just
naturally die.